alpha
My aplogies on the day late. I honestly wanted to go have a good holiday being an Irish pagan. Bit of historical relevance of this. However, when I revisited it this evening. I recognized that the notebook was truncating and it wasn't my reference passing like I thought! So I was able to backtrack and change a function for terminal clearing and it suddenly worked.